# p4est interface performance