--- a/benchmark/benchmarks.jl
+++ b/benchmark/benchmarks.jl
@@ -20,6 +20,7 @@ for elixir in [joinpath(examples_dir(), "tree_2d_dgsem", "elixir_advection_exten
                joinpath(@__DIR__, "elixir_2d_euler_vortex_tree.jl"),
                joinpath(@__DIR__, "elixir_2d_euler_vortex_structured.jl"),
                joinpath(@__DIR__, "elixir_2d_euler_vortex_unstructured.jl"),
+               joinpath(@__DIR__, "elixir_2d_euler_vortex_p4est.jl"),
                joinpath(examples_dir(), "tree_3d_dgsem", "elixir_advection_extended.jl"),
                joinpath(examples_dir(), "tree_3d_dgsem", "elixir_euler_ec.jl"),
                joinpath(examples_dir(), "tree_3d_dgsem", "elixir_euler_mortar.jl"),