This PR improves the performance of surface stuff for the `P4estMesh`. It's a first step with a new design. Instead of using the clean but inefficient code based on `evaluate_index` in inner loops, the logic is moved to a higher level and some loop code is repeated with minor adaptations. This is somehow less clean but also much more efficient. Here are the timings based on one of the setups described in issue #642 (using `julia --check-bounds=no --threads=1`).
```julia
julia> trixi_include(joinpath(examples_dir(), "tree_2d_dgsem", "elixir_euler_vortex.jl"),
           save_restart=TrivialCallback(), save_solution=TrivialCallback())
[...]
 ───────────────────────────────────────────────────────────────────────────────
            Trixi.jl                    Time                   Allocations      
                                ──────────────────────   ───────────────────────
        Tot / % measured:            354ms / 88.7%           1.37MiB / 84.1%    

 Section                ncalls     time   %tot     avg     alloc   %tot      avg
 ───────────────────────────────────────────────────────────────────────────────
 rhs!                    2.78k    288ms  91.9%   104μs   0.99MiB  85.4%     372B
   interface flux        2.78k    119ms  37.9%  42.8μs     0.00B  0.00%    0.00B
   volume integral       2.78k   72.5ms  23.1%  26.1μs     0.00B  0.00%    0.00B
   ~rhs!~                2.78k   44.2ms  14.1%  15.9μs   0.99MiB  85.4%     372B
   surface integral      2.78k   23.6ms  7.51%  8.48μs     0.00B  0.00%    0.00B
   prolong2interfaces    2.78k   16.1ms  5.12%  5.78μs     0.00B  0.00%    0.00B
   reset ∂u/∂t           2.78k   6.95ms  2.22%  2.50μs     0.00B  0.00%    0.00B
   Jacobian              2.78k   5.49ms  1.75%  1.97μs     0.00B  0.00%    0.00B
   prolong2boundaries    2.78k    139μs  0.04%  49.9ns     0.00B  0.00%    0.00B
   prolong2mortars       2.78k    132μs  0.04%  47.3ns     0.00B  0.00%    0.00B
   mortar flux           2.78k   82.6μs  0.03%  29.7ns     0.00B  0.00%    0.00B
   boundary flux         2.78k   79.3μs  0.03%  28.5ns     0.00B  0.00%    0.00B
   source terms          2.78k   59.1μs  0.02%  21.2ns     0.00B  0.00%    0.00B
 calculate dt              557   14.9ms  4.74%  26.7μs     0.00B  0.00%    0.00B
 analyze solution            7   10.5ms  3.36%  1.51ms    173KiB  14.6%  24.7KiB
 ───────────────────────────────────────────────────────────────────────────────

julia> trixi_include(joinpath(examples_dir(), "tree_2d_dgsem", "elixir_euler_vortex.jl"),
           save_restart=TrivialCallback(), save_solution=TrivialCallback(),
           mesh=StructuredMesh((16, 16), float.(coordinates_min), float.(coordinates_max)))
[...]
 ─────────────────────────────────────────────────────────────────────────────
           Trixi.jl                   Time                   Allocations      
                              ──────────────────────   ───────────────────────
       Tot / % measured:           374ms / 88.4%           1.73MiB / 87.2%    

 Section              ncalls     time   %tot     avg     alloc   %tot      avg
 ─────────────────────────────────────────────────────────────────────────────
 rhs!                  2.78k    300ms  90.7%   108μs   0.98MiB  65.3%     370B
   interface flux      2.78k    133ms  40.1%  47.7μs     0.00B  0.00%    0.00B
   volume integral     2.78k   94.3ms  28.5%  33.9μs     0.00B  0.00%    0.00B
   ~rhs!~              2.78k   32.4ms  9.79%  11.7μs   0.98MiB  65.3%     370B
   surface integral    2.78k   25.5ms  7.71%  9.18μs     0.00B  0.00%    0.00B
   Jacobian            2.78k   7.82ms  2.36%  2.81μs     0.00B  0.00%    0.00B
   reset ∂u/∂t         2.78k   7.21ms  2.18%  2.59μs     0.00B  0.00%    0.00B
   boundary flux       2.78k   73.7μs  0.02%  26.5ns     0.00B  0.00%    0.00B
   source terms        2.78k   58.6μs  0.02%  21.1ns     0.00B  0.00%    0.00B
 calculate dt            557   20.1ms  6.08%  36.1μs     0.00B  0.00%    0.00B
 analyze solution          7   10.7ms  3.25%  1.53ms    535KiB  34.7%  76.4KiB
 ─────────────────────────────────────────────────────────────────────────────

julia> trixi_include(joinpath(examples_dir(), "tree_2d_dgsem", "elixir_euler_vortex.jl"),
           save_restart=TrivialCallback(), save_solution=TrivialCallback(),
           mesh=P4estMesh((1, 1), polydeg=3,
                   coordinates_min=coordinates_min, coordinates_max=coordinates_max,
                   initial_refinement_level=4))
[...]
 ───────────────────────────────────────────────────────────────────────────────
            Trixi.jl                    Time                   Allocations      
                                ──────────────────────   ───────────────────────
        Tot / % measured:            804ms / 94.6%           1.73MiB / 87.3%    

 Section                ncalls     time   %tot     avg     alloc   %tot      avg
 ───────────────────────────────────────────────────────────────────────────────
 rhs!                    2.78k    727ms  95.6%   262μs   0.99MiB  65.2%     372B
   prolong2interfaces    2.78k    275ms  36.1%  98.8μs     0.00B  0.00%    0.00B
   interface flux        2.78k    268ms  35.2%  96.2μs     0.00B  0.00%    0.00B
   volume integral       2.78k   96.5ms  12.7%  34.7μs     0.00B  0.00%    0.00B
   ~rhs!~                2.78k   48.9ms  6.43%  17.6μs   0.99MiB  65.2%     372B
   surface integral      2.78k   24.3ms  3.20%  8.74μs     0.00B  0.00%    0.00B
   Jacobian              2.78k   7.59ms  1.00%  2.73μs     0.00B  0.00%    0.00B
   reset ∂u/∂t           2.78k   7.17ms  0.94%  2.58μs     0.00B  0.00%    0.00B
   prolong2mortars       2.78k    157μs  0.02%  56.4ns     0.00B  0.00%    0.00B
   prolong2boundaries    2.78k    147μs  0.02%  52.7ns     0.00B  0.00%    0.00B
   mortar flux           2.78k   89.1μs  0.01%  32.0ns     0.00B  0.00%    0.00B
   boundary flux         2.78k   77.1μs  0.01%  27.7ns     0.00B  0.00%    0.00B
   source terms          2.78k   61.4μs  0.01%  22.1ns     0.00B  0.00%    0.00B
 calculate dt              557   20.8ms  2.73%  37.3μs     0.00B  0.00%    0.00B
 analyze solution            7   12.8ms  1.68%  1.82ms    539KiB  34.8%  77.0KiB
 ───────────────────────────────────────────────────────────────────────────────
```
With the first (verbose) version of PR, I get
```julia
 ───────────────────────────────────────────────────────────────────────────────
            Trixi.jl                    Time                   Allocations      
                                ──────────────────────   ───────────────────────
        Tot / % measured:            394ms / 89.5%           1.73MiB / 87.2%    

 Section                ncalls     time   %tot     avg     alloc   %tot      avg
 ───────────────────────────────────────────────────────────────────────────────
 rhs!                    2.78k    322ms  91.3%   116μs   0.99MiB  65.3%     372B
   interface flux        2.78k    125ms  35.6%  45.1μs     0.00B  0.00%    0.00B
   volume integral       2.78k   91.1ms  25.8%  32.8μs     0.00B  0.00%    0.00B
   ~rhs!~                2.78k   44.2ms  12.6%  15.9μs   0.99MiB  65.3%     372B
   prolong2interfaces    2.78k   24.0ms  6.79%  8.61μs     0.00B  0.00%    0.00B
   surface integral      2.78k   22.5ms  6.37%  8.08μs     0.00B  0.00%    0.00B
   Jacobian              2.78k   7.21ms  2.04%  2.59μs     0.00B  0.00%    0.00B
   reset ∂u/∂t           2.78k   7.03ms  2.00%  2.53μs     0.00B  0.00%    0.00B
   prolong2boundaries    2.78k    143μs  0.04%  51.3ns     0.00B  0.00%    0.00B
   prolong2mortars       2.78k    136μs  0.04%  48.9ns     0.00B  0.00%    0.00B
   mortar flux           2.78k   87.8μs  0.02%  31.6ns     0.00B  0.00%    0.00B
   boundary flux         2.78k   78.4μs  0.02%  28.2ns     0.00B  0.00%    0.00B
   source terms          2.78k   61.1μs  0.02%  22.0ns     0.00B  0.00%    0.00B
 calculate dt              557   20.0ms  5.66%  35.8μs     0.00B  0.00%    0.00B
 analyze solution            7   10.6ms  3.01%  1.52ms    536KiB  34.7%  76.6KiB
 ───────────────────────────────────────────────────────────────────────────────
```
With the second version mimicking the handling of interfaces for the `UnstructuredMesh2D`, I get
```
 ───────────────────────────────────────────────────────────────────────────────
            Trixi.jl                    Time                   Allocations      
                                ──────────────────────   ───────────────────────
        Tot / % measured:            417ms / 87.4%           1.73MiB / 87.4%    

 Section                ncalls     time   %tot     avg     alloc   %tot      avg
 ───────────────────────────────────────────────────────────────────────────────
 rhs!                    2.78k    334ms  91.5%   120μs   0.99MiB  65.3%     372B
   interface flux        2.78k    138ms  37.8%  49.6μs     0.00B  0.00%    0.00B
   volume integral       2.78k   87.8ms  24.1%  31.6μs     0.00B  0.00%    0.00B
   ~rhs!~                2.78k   42.1ms  11.5%  15.1μs   0.99MiB  65.3%     372B
   prolong2interfaces    2.78k   29.2ms  8.01%  10.5μs     0.00B  0.00%    0.00B
   surface integral      2.78k   22.6ms  6.19%  8.12μs     0.00B  0.00%    0.00B
   reset ∂u/∂t           2.78k   6.85ms  1.88%  2.46μs     0.00B  0.00%    0.00B
   Jacobian              2.78k   6.78ms  1.86%  2.44μs     0.00B  0.00%    0.00B
   prolong2boundaries    2.78k    139μs  0.04%  50.2ns     0.00B  0.00%    0.00B
   prolong2mortars       2.78k    119μs  0.03%  42.6ns     0.00B  0.00%    0.00B
   mortar flux           2.78k   93.7μs  0.03%  33.7ns     0.00B  0.00%    0.00B
   boundary flux         2.78k   71.8μs  0.02%  25.8ns     0.00B  0.00%    0.00B
   source terms          2.78k   46.6μs  0.01%  16.8ns     0.00B  0.00%    0.00B
 calculate dt              557   19.7ms  5.41%  35.4μs     0.00B  0.00%    0.00B
 analyze solution            7   11.4ms  3.11%  1.62ms    538KiB  34.7%  76.8KiB
 ───────────────────────────────────────────────────────────────────────────────
```
That's ca. 10% less efficient than the verbose first version but still  much more efficient than the version in `main`.

The current version is limited to interfaces in 2D. Further changes that would need to be addressed if we follow this design are
- [x] 2D interfaces
- [x] 2D boundaries
- [x] 2D mortars

Left for another PR later (listed in #642):
- 3D interfaces
- 3D boundaries
- 3D mortars
- Look for `TODO: p4est interface performance`

I would like to get your feedback on this approach at first such that we can discuss whether there is a better approach. If not, I suggest to merge this (since it's easy enough to review) and perform more performance improvements in other PRs step by step to attack #642.