@ranocha 2021-08-05
Thanks you very much for your input, @efaulhaber. 

> However, I'm still not convinced that pulling the `if` out of this loop is the best solution, and that there is no way to let the compiler handle this part. If this is really the only way (for now at least), I'm fine with it for 2D.

I'm open to better suggestions but I haven't found one yet... Do you have an idea?

> But I guess the 3D code will be horribly ugly, as we will end up with a lot of boilerplate code there. Wouldn't that be 48 possibilities, in both interface and mortar logic? Plus the 12 for the primary element?

I didn't count the number of possibilities but that's definitely the bad part about it... How do you usually handle this, @andrewwinters5000?