@efaulhaber 2021-08-05
> I'm open to better suggestions but I haven't found one yet... Do you have an idea?

Unfortunately, no. The best ideas I had with my limited Julia knowledge are the ones I described in #642.