@ranocha 2021-08-11
We discussed this in the Trixi meeting yesterday with @andrewwinters5000 and decided to make a compromise of performance and readability/simplicity of the code. I implemented such an approach loosely mimicking what we do for the `UnstrucutredMesh2D` at the interfaces. See the updated performance numbers in the first post. If you don't have any better idea, this seems to be an acceptable compromise.