@ranocha 2021-08-12
> Have you run a test what this does for the timings of a simulation with mortar elements?

```julia
julia> using Trixi

julia> trixi_include("examples/p4est_2d_dgsem/elixir_advection_amr_unstructured_flag.jl",
                     save_solution=TrivialCallback(), save_restart=TrivialCallback(),
                     tspan=(0.0, 100.0))
[...]
 Section                     ncalls     time   %tot     avg     alloc   %tot      avg
 ────────────────────────────────────────────────────────────────────────────────────
 rhs!                         5.33k    486ms  62.2%  91.2μs    166MiB  47.4%  31.9KiB
   prolong2mortars            5.33k    158ms  20.2%  29.7μs    164MiB  46.9%  31.5KiB
   interface flux             5.33k   81.3ms  10.4%  15.3μs     0.00B  0.00%    0.00B
   volume integral            5.33k   63.5ms  8.12%  11.9μs     0.00B  0.00%    0.00B
   ~rhs!~                     5.33k   58.1ms  7.43%  10.9μs   1.88MiB  0.54%     370B
   prolong2interfaces         5.33k   49.4ms  6.31%  9.26μs     0.00B  0.00%    0.00B
   boundary flux              5.33k   29.7ms  3.79%  5.56μs     0.00B  0.00%    0.00B
   mortar flux                5.33k   24.1ms  3.08%  4.51μs     0.00B  0.00%    0.00B
   surface integral           5.33k   10.3ms  1.32%  1.93μs     0.00B  0.00%    0.00B
   Jacobian                   5.33k   7.25ms  0.93%  1.36μs     0.00B  0.00%    0.00B
   prolong2boundaries         5.33k   2.36ms  0.30%   443ns     0.00B  0.00%    0.00B
   reset ∂u/∂t                5.33k   1.99ms  0.25%   374ns     0.00B  0.00%    0.00B
   source terms               5.33k    114μs  0.01%  21.3ns     0.00B  0.00%    0.00B
```
This PR:
```julia
julia> using Trixi

julia> trixi_include("examples/p4est_2d_dgsem/elixir_advection_amr_unstructured_flag.jl",
                     save_solution=TrivialCallback(), save_restart=TrivialCallback(),
                     tspan=(0.0, 100.0))
[...]
 Section                     ncalls     time   %tot     avg     alloc   %tot      avg
 ────────────────────────────────────────────────────────────────────────────────────
 rhs!                         5.33k    399ms  55.5%  74.8μs    166MiB  47.4%  31.9KiB
   prolong2mortars            5.33k    160ms  22.3%  30.1μs    164MiB  46.9%  31.5KiB
   ~rhs!~                     5.33k   68.1ms  9.47%  12.8μs   1.88MiB  0.54%     370B
   volume integral            5.33k   62.4ms  8.68%  11.7μs     0.00B  0.00%    0.00B
   interface flux             5.33k   35.1ms  4.88%  6.58μs     0.00B  0.00%    0.00B
   boundary flux              5.33k   25.4ms  3.53%  4.77μs     0.00B  0.00%    0.00B
   prolong2interfaces         5.33k   15.7ms  2.18%  2.95μs     0.00B  0.00%    0.00B
   mortar flux                5.33k   11.4ms  1.58%  2.13μs     0.00B  0.00%    0.00B
   surface integral           5.33k   9.93ms  1.38%  1.86μs     0.00B  0.00%    0.00B
   Jacobian                   5.33k   7.10ms  0.99%  1.33μs     0.00B  0.00%    0.00B
   reset ∂u/∂t                5.33k   1.89ms  0.26%   354ns     0.00B  0.00%    0.00B
   prolong2boundaries         5.33k   1.56ms  0.22%   293ns     0.00B  0.00%    0.00B
   source terms               5.33k    115μs  0.02%  21.7ns     0.00B  0.00%    0.00B
```
So there is still #628